refactor: ResponseDto 위치 이동
