+++ backend/src/main/java/org/example/backend/professor/controller/ProfessorController.java
@@ -4,11 +4,10 @@ package org.example.backend.professor.controller;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.example.backend.professor.domain.dto.ResponseDto;
+import org.example.backend.common.exception.dto.ResponseDto;
 import org.example.backend.professor.domain.dto.professor.ProfessorReqDto;
 import org.example.backend.professor.domain.dto.professor.ProfessorResDto;
 import org.example.backend.professor.service.ProfessorService;
-import org.example.backend.thesis.domain.dto.ThesisResDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
