+++ backend/src/main/java/org/example/backend/thesis/controller/ThesisController.java
@@ -3,8 +3,7 @@ package org.example.backend.thesis.controller;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.example.backend.board.domain.dto.BoardResDto;
-import org.example.backend.professor.domain.dto.ResponseDto;
+import org.example.backend.common.exception.dto.ResponseDto;
 import org.example.backend.thesis.domain.dto.ThesisReqDto;
 import org.example.backend.thesis.domain.dto.ThesisResDto;
 import org.example.backend.thesis.service.ThesisService;
